refactor the components to be in particular folders
