+++ frontend/src/components/usersPage/ConfirmationModal.test.tsx
@@ -1,7 +1,7 @@
 import { fireEvent, render, screen } from '@testing-library/react';
 import { beforeEach, describe, expect, it, vi } from 'vitest';
 
-import { TableData } from '../types';
+import { TableData } from '../../types';
 import { ConfirmationModal } from './ConfirmationModal';
 
 const mockUser: TableData = {
