+++ frontend/src/contexts/ToastContext.tsx
@@ -1,6 +1,6 @@
 import React, { createContext, useContext } from 'react';
 
-import { ToastContainer } from '../components/ToastContainer';
+import { ToastContainer } from '../components';
 import { useToast } from '../hooks/useToast';
 
 interface ToastContextType {
