+++ frontend/src/router/AppRouter.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 
-import { Layout } from '../components/Layout';
+import { Layout } from '../components';
 import { FeaturesPage } from '../pages/help/FeaturesPage';
 import { GettingStartedPage } from '../pages/help/GettingStartedPage';
 import { UserManagementPage } from '../pages/help/UserManagementPage';
